Reduce the blur and spread radius size of a reftest.
In the near future, we'll be changing the blur algorithm in WR.

On native GPUs, this is fine - but on OSMesa, a blur of this
radius and spread is quite slow, and can cause test timeouts
when running with a lot of processes enabled.

Reducing the spread / blur shouldn't affect what this is testing.