Fix mutation test in CI by passing env vars correctly
Originally, buildbot would try to directly execute `RUSTFLAGS=` as a binary
since it was the first token in the command string, which won't work.

Fix this by using the `env` command to set the environment variables,
as there is currently no support for setting per-step environment variables
in the `buildbot_steps.yml` file.