Fix `./mach filter_intermittents` in shallow git clones.
`git log --merges` only shows commits with at least two parents.
Presumably, this only works when the parent commits are in the local repository.
In a shallow clone (`--depth 1`), on the HEAD commit is fetched.

This commit removes the `--merges` filter.
This shouldn’t affect the behavior,
since CI always runs with a merge commit as HEAD.