compositor: Limit animation tick interval when in WPT mode (#31561)
When exiting after load or dumping a screenshot, limit the speed that
the compositor sends tick events to the constellation (and on to the
script thread). Previously, these tick messages were sent as fast as the
winit event loop (that's really fast!), until the document was ready to
for its screenshot. rAf is meant to run at more or less the same
frequency as the refresh rate of the screen.

On my system, the massive amount of ticks was overwhelming the router
thread, preventing fonts from loading properly. This change is an attempt
to make tests with requestAnimationFrame less flaky.