wpt: Fix `/css/CSS2/tables/anonymous-table-box-width-001.xht` (#35162)
This test was failing on all browsers because before the paragraph
"Test passes if there is a filled green square and *no red*" was getting
styled with some unnecessary CSS which is not present in the reference.

Removing this CSS which is irrelevant to the actual table being tested
fixes the problem.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>