Rename ohos app bundle (#35790)
The `shell` suffix causes issues with the startup profiling tools on
OpenHarmony. The profiler fails to detect the start of the app.
This is fixed by renaming the app. I tried various
different variations and identified the `shell`
suffix to be the culprit.
E.g. `org.servo.shell` has the same issue, so its not
a length issue.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>