layout: Remove `calculate_hypothetical_cross_size()` (#35821)
One of the callers was only used for an assert, and it was passing an
inline size argument to a parameter expecting a block size, so it wasn't
making much sense anyways.

Just inline the code into the other caller, and for consistency remove
the assert for replaced elements too.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>